ClaudeTrading/requirements.txt
alpaca-py>=0.20.0
pandas>=2.0.0
requests>=2.31.0
schedule>=1.2.0
